requirements.txt
av
einops
ffmpeg
fvcore
imageio
imageio[ffmpeg]
joblib
kubric
lovely_numpy
lovely_tensors
matplotlib
numpy
opencv_python
pandas
Pillow
Pillow
psutil
pybullet
rich
scikit_learn
scipy
seaborn
setuptools
simplejson
timm
torch_optimizer
tqdm
wandb
